requirements.txt
aiodns==2.0.0
aiofile==3.5.0
aiohttp==3.7.4.post0
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.5
boto3==1.17.28
botocore==1.20.28
brotlipy==0.7.0
caio==0.8.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coverage==5.5
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
django-appconf==1.0.4
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-environ==0.4.5
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-imagekit==4.0.2
django-redis==4.12.1
django-storages==1.11.1
django-upload-validator==1.1.5
docutils==0.16
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jmespath==0.10.0
libmagic==1.0
msgpack==1.0.2
multidict==5.1.0
pilkit==2.0
Pillow==8.1.2
promise==2.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-magic==0.4.22
pytz==2021.1
redis==3.5.3
requests==2.25.1
Rx==1.6.1
s3transfer==0.3.4
service-identity==18.1.0
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
zope.interface==5.2.0
